backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
aiofiles>=23.2.1
pydantic>=2.0.0

torch>=2.1.2 --index-url https://download.pytorch.org/whl/cu121
torchaudio>=2.1.2 --index-url https://download.pytorch.org/whl/cu121

transformers>=4.40.0
sentence-transformers>=2.7.0
accelerate>=0.30.0
huggingface_hub>=0.23.0

pyannote.audio>=3.1.1
soundfile>=0.12.1
ffmpeg-python>=0.2.0

faiss-cpu>=1.7.4

scikit-learn>=1.4.0
numpy>=1.26.0
hdbscan>=0.8.33         

assemblyai>=0.23.0

scipy>=1.11.0